Added a MessageDispatcher class with ICommunicator and IRemoteIdentifier for layering sockets Added TickableObjects and ITickable, which replaces NetworkManager and ITickableNetwork Modified RequestHandler so that it fits message dispatcher layer
Assets :
- Replaced Load() bool return value by void, instead use exceptions

Network/Client/CGModule :
- Now parse serverType
- Stores mapChecksum and votes
- Increased maximum number of objectives to 5

Network/Client/Server | Network/Client/MasterServer | Network/Client/ClientGame:
- Using MessageDispatcher instead of one global NetworkManager

Network/Types:
- Added Spearhead serverType_e

Managers/ShaderManager:
- Added support for more shader functions

MOHPC::Log::logType_e: now using common enum
Lot of minor fixes

+++ mohpc-lib/Formats/BSP/BSP_Curve.cpp
@@ -280,7 +280,7 @@ void BSP::CreateSurfaceGridMesh(int32_t width, int32_t height, BSPData::Vertice 
 	grid->indexes.resize(numIndexes);
 	grid->vertices.resize(width * height);
 
-	size_t *outIndexes = grid->indexes.data();
+	uint32_t *outIndexes = grid->indexes.data();
 	for (size_t i = 0; i < numIndexes; i++)
 	{
 		outIndexes[i] = indexes[i];
